Added 3 files, Modified 2 files
- Added files: code/chain_server/chat_templates.py,
code/chatui/pages/info.py, code/chatui/pages/utils.py
- Modified files: code/chain_server/chains.py,
code/chatui/pages/converse.py
