+++ code/chain_server/chains.py
@@ -62,103 +62,14 @@ if TYPE_CHECKING:
 
     from chain_server.configuration_wizard import ConfigWizard
 
+from chain_server import chat_templates
+
 TEXT_SPLITTER_MODEL = "intfloat/e5-large-v2"
 TEXT_SPLITTER_CHUNCK_SIZE = 510
 TEXT_SPLITTER_CHUNCK_OVERLAP = 200
 EMBEDDING_MODEL = "intfloat/e5-large-v2"
 DEFAULT_NUM_TOKENS = 50
 DEFAULT_MAX_CONTEXT = 800
-
-MISTRAL_CHAT_TEMPLATE = (
-    "<s>[INST]"
-    "You are a helpful, respectful and honest assistant. Always answer as helpfully as possible, while being safe. Your answers should not include any harmful, unethical, dangerous, or illegal content. Please ensure that your responses are positive in nature."
-    "If a question does not make any sense, or is not factually coherent, explain why instead of answering something not correct. If you don't know the answer to a question, please don't share false information."
-    "[/INST] {context_str} </s> [INST] {query_str} [/INST]"
-)
-
-LLAMA_2_CHAT_TEMPLATE = (
-    "<s>[INST] <<SYS>>"
-    "You are a helpful, respectful and honest assistant. Always answer as helpfully as possible, while being safe. Your answers should not include any harmful, unethical, dangerous, or illegal content. Please ensure that your responses are positive in nature."
-    "If a question does not make any sense, or is not factually coherent, explain why instead of answering something not correct. If you don't know the answer to a question, please don't share false information."
-    "<</SYS>>"
-    "[/INST] {context_str} </s><s>[INST] {query_str} [/INST]"
-)
-
-LLAMA_3_CHAT_TEMPLATE = (
-    "<|begin_of_text|><|start_header_id|>system<|end_header_id|>"
-    "You are a helpful, respectful and honest assistant. Always answer as helpfully as possible, while being safe. Your answers should not include any harmful, unethical, dangerous, or illegal content. Please ensure that your responses are positive in nature."
-    "If a question does not make any sense, or is not factually coherent, explain why instead of answering something not correct. If you don't know the answer to a question, please don't share false information."
-    "<|eot_id|><|start_header_id|>user<|end_header_id|>\n"
-    "{context_str} {query_str}<|eot_id|><|start_header_id|>assistant<|end_header_id|>"
-)
-
-NVIDIA_CHAT_TEMPLATE = (
-    "System: This is a chat between a user and an artificial intelligence assistant. The assistant gives helpful, detailed, and polite answers to the user's questions based on the context. The assistant should also indicate when the answer cannot be found in the context. \n"
-    "User: {context_str} {query_str} \n"
-    "Assistant: "
-)
-
-MICROSOFT_CHAT_TEMPLATE = (
-    "<|user|>\n"
-    "You are a helpful, respectful and honest assistant. Always answer as helpfully as possible, while being safe. Your answers should not include any harmful, unethical, dangerous, or illegal content. If you don't know the answer to a question, please don't share false information. Please ensure that your responses are positive in nature.\n"
-    "The user's question is: {context_str} {query_str} <|end|> \n"
-    "<|assistant|>"
-)
-
-GENERIC_CHAT_TEMPLATE = (
-    "You are a helpful, respectful and honest assistant. Always answer as helpfully as possible, while being safe. Your answers should not include any harmful, unethical, dangerous, or illegal content. If you don't know the answer to a question, please don't share false information. Please ensure that your responses are positive in nature.\n"
-    "The user's question is: {context_str} {query_str} <|end|> \n"
-)
-
-MISTRAL_RAG_TEMPLATE = (
-    "<s>[INST] <<SYS>>"
-    "Use the following context to answer the user's question. If you don't know the answer,"
-    "just say that you don't know, don't try to make up an answer."
-    "<</SYS>>"
-    "<s>[INST] Context: {context_str} Question: {query_str} Only return the helpful"
-    " answer below and nothing else. Helpful answer:[/INST]"
-)
-
-LLAMA_2_RAG_TEMPLATE = (
-    "<s>[INST] <<SYS>>"
-    "Use the following context to answer the user's question. If you don't know the answer,"
-    "just say that you don't know, don't try to make up an answer."
-    "<</SYS>>"
-    "<s>[INST] Context: {context_str} Question: {query_str} Only return the helpful"
-    " answer below and nothing else. Helpful answer:[/INST]"
-)
-
-LLAMA_3_RAG_TEMPLATE = (
-    "<|begin_of_text|><|start_header_id|>system<|end_header_id|>"
-    "Use the following context to answer the user's question. If you don't know the answer,"
-    "just say that you don't know, don't try to make up an answer."
-    "<|eot_id|><|start_header_id|>user<|end_header_id|>\n"
-    "Context: {context_str} Question: {query_str} Only return the helpful"
-    " answer below and nothing else. Helpful answer:<|eot_id|><|start_header_id|>assistant<|end_header_id|>"
-)
-
-NVIDIA_RAG_TEMPLATE = (
-    "System: This is a chat between a user and an artificial intelligence assistant. The assistant gives helpful, detailed, and polite answers to the user's questions based on the context. The assistant should also indicate when the answer cannot be found in the context.\n"
-    "{context_str} \n"
-    "User: {query_str} \n"
-    "Assistant: "
-)
-
-MICROSOFT_RAG_TEMPLATE = (
-    "<|user|>\n"
-    "Use the following context to answer the question. If you don't know the answer,"
-    "just say that you don't know, don't try to make up an answer.\n"
-    "Context: {context_str} Question: {query_str} Only return the helpful"
-    " answer below and nothing else. <|end|> \n" 
-    "<|assistant|>"
-)
-
-GENERIC_RAG_TEMPLATE = (
-    "Use the following context to answer the question. If you don't know the answer,"
-    "just say that you don't know, don't try to make up an answer.\n"
-    "Context: {context_str} Question: {query_str} Only return the helpful"
-    " answer below and nothing else. \n" 
-)
 
 
 class LimitRetrievedNodesLength(BaseNodePostprocessor):
@@ -292,17 +203,17 @@ def llm_chain_streaming(
 
     if inference_mode == "local":
         if "nvidia" in local_model_id:
-            prompt = NVIDIA_CHAT_TEMPLATE.format(context_str=context, query_str=question)
+            prompt = chat_templates.NVIDIA_CHAT_TEMPLATE.format(context_str=context, query_str=question)
         elif "Llama-3" in local_model_id:
-            prompt = LLAMA_3_CHAT_TEMPLATE.format(context_str=context, query_str=question)
+            prompt = chat_templates.LLAMA_3_CHAT_TEMPLATE.format(context_str=context, query_str=question)
         elif "Llama-2" in local_model_id:
-            prompt = LLAMA_2_CHAT_TEMPLATE.format(context_str=context, query_str=question)
+            prompt = chat_templates.LLAMA_2_CHAT_TEMPLATE.format(context_str=context, query_str=question)
         elif "microsoft" in local_model_id:
-            prompt = MICROSOFT_CHAT_TEMPLATE.format(context_str=context, query_str=question)
+            prompt = chat_templates.MICROSOFT_CHAT_TEMPLATE.format(context_str=context, query_str=question)
         elif "mistralai" in local_model_id:
-            prompt = MISTRAL_CHAT_TEMPLATE.format(context_str=context, query_str=question)
+            prompt = chat_templates.MISTRAL_CHAT_TEMPLATE.format(context_str=context, query_str=question)
         else: 
-            prompt = NVIDIA_CHAT_TEMPLATE.format(context_str=context, query_str=question)
+            prompt = chat_templates.NVIDIA_CHAT_TEMPLATE.format(context_str=context, query_str=question)
         start = time.time()
         response = get_llm(inference_mode, nvcf_model_id, nim_model_ip, num_tokens, temp, top_p, freq_pen).stream_complete(prompt, max_new_tokens=num_tokens)
         perf = time.time() - start
@@ -315,15 +226,15 @@ def llm_chain_streaming(
                 break
     else:
         if inference_mode == "cloud" and "llama3" in nvcf_model_id:
-            prompt = LLAMA_3_CHAT_TEMPLATE.format(context_str=context, query_str=question)
+            prompt = chat_templates.LLAMA_3_CHAT_TEMPLATE.format(context_str=context, query_str=question)
         elif inference_mode == "cloud" and "llama2" in nvcf_model_id:
-            prompt = LLAMA_2_CHAT_TEMPLATE.format(context_str=context, query_str=question)
+            prompt = chat_templates.LLAMA_2_CHAT_TEMPLATE.format(context_str=context, query_str=question)
         elif inference_mode == "cloud" and "mistral" in nvcf_model_id:
-            prompt = MISTRAL_CHAT_TEMPLATE.format(context_str=context, query_str=question)
+            prompt = chat_templates.MISTRAL_CHAT_TEMPLATE.format(context_str=context, query_str=question)
         elif inference_mode == "cloud" and "microsoft" in nvcf_model_id:
-            prompt = MICROSOFT_CHAT_TEMPLATE.format(context_str=context, query_str=question)
+            prompt = chat_templates.MICROSOFT_CHAT_TEMPLATE.format(context_str=context, query_str=question)
         else:
-            prompt = GENERIC_CHAT_TEMPLATE.format(context_str=context, query_str=question)
+            prompt = chat_templates.GENERIC_CHAT_TEMPLATE.format(context_str=context, query_str=question)
         openai.api_key = os.environ.get('NVCF_RUN_KEY') if inference_mode == "cloud" else "xyz"
         openai.base_url = "https://integrate.api.nvidia.com/v1/" if inference_mode == "cloud" else "http://" + nim_model_ip + ":" + ("8000" if len(nim_model_port) == 0 else nim_model_port) + "/v1/"
 
@@ -379,17 +290,17 @@ def rag_chain_streaming(prompt: str,
         for node in nodes: 
             docs.append(node.get_text())
         if "nvidia" in local_model_id:
-            prompt = NVIDIA_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
+            prompt = chat_templates.NVIDIA_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
         elif "Llama-3" in local_model_id:
-            prompt = LLAMA_3_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
+            prompt = chat_templates.LLAMA_3_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
         elif "Llama-2" in local_model_id:
-            prompt = LLAMA_2_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
+            prompt = chat_templates.LLAMA_2_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
         elif "microsoft" in local_model_id:
-            prompt = MICROSOFT_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
+            prompt = chat_templates.MICROSOFT_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
         elif "mistralai" in local_model_id:
-            prompt = MISTRAL_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
+            prompt = chat_templates.MISTRAL_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
         else: 
-            prompt = NVIDIA_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
+            prompt = chat_templates.NVIDIA_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
         start = time.time()
         response = get_llm(inference_mode, 
                            nvcf_model_id, 
@@ -416,15 +327,15 @@ def rag_chain_streaming(prompt: str,
         for node in nodes: 
             docs.append(node.get_text())
         if inference_mode == "cloud" and "llama3" in nvcf_model_id:
-            prompt = LLAMA_3_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
+            prompt = chat_templates.LLAMA_3_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
         elif inference_mode == "cloud" and "llama2" in nvcf_model_id:
-            prompt = LLAMA_2_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
+            prompt = chat_templates.LLAMA_2_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
         elif inference_mode == "cloud" and "mistral" in nvcf_model_id:
-            prompt = MISTRAL_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
+            prompt = chat_templates.MISTRAL_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
         elif inference_mode == "cloud" and "microsoft" in nvcf_model_id:
-            prompt = MICROSOFT_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
+            prompt = chat_templates.MICROSOFT_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
         else:
-            prompt = GENERIC_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
+            prompt = chat_templates.GENERIC_RAG_TEMPLATE.format(context_str=", ".join(docs), query_str=prompt)
         start = time.time()
         completion = openai.chat.completions.create(
           model= nvcf_model_id if inference_mode == "cloud" else ("meta/llama3-8b-instruct" if len(nim_model_id) == 0 else nim_model_id),
