+++ code/chain_server/chat_templates.py
@@ -1,0 +1,107 @@
+# SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""LLM Chat Templates for executing Retrival Augmented Generation."""
+
+MISTRAL_CHAT_TEMPLATE = (
+    "<s>[INST]"
+    "You are a helpful, respectful and honest assistant. Always answer as helpfully as possible, while being safe. Your answers should not include any harmful, unethical, dangerous, or illegal content. Please ensure that your responses are positive in nature."
+    "If a question does not make any sense, or is not factually coherent, explain why instead of answering something not correct. If you don't know the answer to a question, please don't share false information."
+    "[/INST] {context_str} </s> [INST] {query_str} [/INST]"
+)
+
+LLAMA_2_CHAT_TEMPLATE = (
+    "<s>[INST] <<SYS>>"
+    "You are a helpful, respectful and honest assistant. Always answer as helpfully as possible, while being safe. Your answers should not include any harmful, unethical, dangerous, or illegal content. Please ensure that your responses are positive in nature."
+    "If a question does not make any sense, or is not factually coherent, explain why instead of answering something not correct. If you don't know the answer to a question, please don't share false information."
+    "<</SYS>>"
+    "[/INST] {context_str} </s><s>[INST] {query_str} [/INST]"
+)
+
+LLAMA_3_CHAT_TEMPLATE = (
+    "<|begin_of_text|><|start_header_id|>system<|end_header_id|>"
+    "You are a helpful, respectful and honest assistant. Always answer as helpfully as possible, while being safe. Your answers should not include any harmful, unethical, dangerous, or illegal content. Please ensure that your responses are positive in nature."
+    "If a question does not make any sense, or is not factually coherent, explain why instead of answering something not correct. If you don't know the answer to a question, please don't share false information."
+    "<|eot_id|><|start_header_id|>user<|end_header_id|>\n"
+    "{context_str} {query_str}<|eot_id|><|start_header_id|>assistant<|end_header_id|>"
+)
+
+NVIDIA_CHAT_TEMPLATE = (
+    "System: This is a chat between a user and an artificial intelligence assistant. The assistant gives helpful, detailed, and polite answers to the user's questions based on the context. The assistant should also indicate when the answer cannot be found in the context. \n"
+    "User: {context_str} {query_str} \n"
+    "Assistant: "
+)
+
+MICROSOFT_CHAT_TEMPLATE = (
+    "<|user|>\n"
+    "You are a helpful, respectful and honest assistant. Always answer as helpfully as possible, while being safe. Your answers should not include any harmful, unethical, dangerous, or illegal content. If you don't know the answer to a question, please don't share false information. Please ensure that your responses are positive in nature.\n"
+    "The user's question is: {context_str} {query_str} <|end|> \n"
+    "<|assistant|>"
+)
+
+GENERIC_CHAT_TEMPLATE = (
+    "You are a helpful, respectful and honest assistant. Always answer as helpfully as possible, while being safe. Your answers should not include any harmful, unethical, dangerous, or illegal content. If you don't know the answer to a question, please don't share false information. Please ensure that your responses are positive in nature.\n"
+    "The user's question is: {context_str} {query_str} <|end|> \n"
+)
+
+MISTRAL_RAG_TEMPLATE = (
+    "<s>[INST] <<SYS>>"
+    "Use the following context to answer the user's question. If you don't know the answer,"
+    "just say that you don't know, don't try to make up an answer."
+    "<</SYS>>"
+    "<s>[INST] Context: {context_str} Question: {query_str} Only return the helpful"
+    " answer below and nothing else. Helpful answer:[/INST]"
+)
+
+LLAMA_2_RAG_TEMPLATE = (
+    "<s>[INST] <<SYS>>"
+    "Use the following context to answer the user's question. If you don't know the answer,"
+    "just say that you don't know, don't try to make up an answer."
+    "<</SYS>>"
+    "<s>[INST] Context: {context_str} Question: {query_str} Only return the helpful"
+    " answer below and nothing else. Helpful answer:[/INST]"
+)
+
+LLAMA_3_RAG_TEMPLATE = (
+    "<|begin_of_text|><|start_header_id|>system<|end_header_id|>"
+    "Use the following context to answer the user's question. If you don't know the answer,"
+    "just say that you don't know, don't try to make up an answer."
+    "<|eot_id|><|start_header_id|>user<|end_header_id|>\n"
+    "Context: {context_str} Question: {query_str} Only return the helpful"
+    " answer below and nothing else. Helpful answer:<|eot_id|><|start_header_id|>assistant<|end_header_id|>"
+)
+
+NVIDIA_RAG_TEMPLATE = (
+    "System: This is a chat between a user and an artificial intelligence assistant. The assistant gives helpful, detailed, and polite answers to the user's questions based on the context. The assistant should also indicate when the answer cannot be found in the context.\n"
+    "{context_str} \n"
+    "User: {query_str} \n"
+    "Assistant: "
+)
+
+MICROSOFT_RAG_TEMPLATE = (
+    "<|user|>\n"
+    "Use the following context to answer the question. If you don't know the answer,"
+    "just say that you don't know, don't try to make up an answer.\n"
+    "Context: {context_str} Question: {query_str} Only return the helpful"
+    " answer below and nothing else. <|end|> \n" 
+    "<|assistant|>"
+)
+
+GENERIC_RAG_TEMPLATE = (
+    "Use the following context to answer the question. If you don't know the answer,"
+    "just say that you don't know, don't try to make up an answer.\n"
+    "Context: {context_str} Question: {query_str} Only return the helpful"
+    " answer below and nothing else. \n" 
+)
